systhreads: extract common code that is not posix/pthreads
