+++ otherlibs/systhreads/st_pthreads.h
@@ -26,20 +26,6 @@
 #include <unistd.h>
 #endif
 
-typedef int st_retcode;
-
-/* Variables used to stop "tick" threads */
-static atomic_uintnat tick_thread_stop[Max_domains];
-#define Tick_thread_stop tick_thread_stop[Caml_state->id]
-
-/* OS-specific initialization */
-
-static int st_initialize(void)
-{
-  atomic_store_release(&Tick_thread_stop, 0);
-  return 0;
-}
-
 /* Thread creation.  Created in detached mode if [res] is NULL. */
 
 typedef pthread_t st_thread_id;
@@ -93,14 +79,14 @@ Caml_inline void st_tls_set(st_tlskey k, void * v)
    all risks of busy-waiting.  Also, we count the number of waiting
    threads. */
 
-typedef struct {
+struct st_masterlock {
   int init;                       /* have the mutex and the cond been
                                      initialized already? */
   pthread_mutex_t lock;           /* to protect contents */
   uintnat busy;                   /* 0 = free, 1 = taken */
   atomic_uintnat waiters;         /* number of threads waiting on master lock */
   pthread_cond_t is_free;         /* signaled when free */
-} st_masterlock;
+};
 
 /* Returns non-zero on failure */
 static int st_masterlock_init(st_masterlock * m)
@@ -131,40 +117,6 @@ static void st_masterlock_destroy(st_masterlock * m)
   rc = pthread_mutex_destroy(&m->lock);
   CAMLassert(!rc);
   (void)rc;
-}
-
-static uintnat st_masterlock_waiters(st_masterlock * m)
-{
-  return atomic_load_acquire(&m->waiters);
-}
-
-static void st_bt_lock_acquire(st_masterlock *m) {
-
-  /* We do not want to signal the backup thread if it is not "working"
-     as it may very well not be, because we could have just resumed
-     execution from another thread right away. */
-  if (caml_bt_is_in_blocking_section()) {
-    caml_bt_enter_ocaml();
-  }
-
-  caml_acquire_domain_lock();
-
-  return;
-}
-
-static void st_bt_lock_release(st_masterlock *m) {
-
-  /* Here we do want to signal the backup thread iff there's
-     no thread waiting to be scheduled, and the backup thread is currently
-     idle. */
-  if (st_masterlock_waiters(m) == 0 &&
-      caml_bt_is_in_blocking_section() == 0) {
-    caml_bt_exit_ocaml();
-  }
-
-  caml_release_domain_lock();
-
-  return;
 }
 
 static void st_masterlock_acquire(st_masterlock *m)
@@ -300,22 +252,4 @@ static int st_event_wait(st_event e)
   }
   rc = pthread_mutex_unlock(&e->lock);
   return rc;
-}
-
-/* The tick thread: interrupt the domain periodically to force preemption  */
-
-static void * caml_thread_tick(void * arg)
-{
-  int *domain_id = (int *) arg;
-
-  caml_init_domain_self(*domain_id);
-  caml_domain_state *domain = Caml_state;
-
-  while(! atomic_load_acquire(&Tick_thread_stop)) {
-    st_msleep(Thread_timeout);
-
-    atomic_store_release(&domain->requested_external_interrupt, 1);
-    caml_interrupt_self();
-  }
-  return NULL;
 }
